Move heartbeat code to APNsClient class
